Various Skills, Stats and Bugfixes
- New Skill: Buckshot
Fire a high caliber round that knocks you and the target enemy away. The projectile explodes on impact, and leaves a trail of fire behind the target.
- New Skill: Mark Target
Marks the target, causing it to receive more physical damage. Rifle-Skills can have special interactions with marked enemies.
- Adjusted Skill: Reload
Now slows, mutes and disarmes you for the duration of the reload time.